Various Updates
added cooldown to dash
added cooldown to melee weapon
Began work on stat level ups in code (works in the inspector rn no way to add in game loop atm)
made hp be able to level up
made mana be able to level up
added damage reduction based on def level
added damage multiplier based on atk level
Made bullets able to play respective firing sounds
added rotary menu UI
added temp assets for aim reticle and the waypoint
made the colour of the aim reticle corrospond to the bullet equiped

+++ Assets/Scripts/Gameplay Scipts/EnemyScripts/EnemySpawner.cs
@@ -10,7 +10,7 @@ public class EnemySpawner : MonoBehaviour
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     public GameObject enemyBase;
     public EnemyBase[] enemyTypes;
-    public int HP;
+    public float HP;
     bool dead = false;
     public bool temporary;
     private int spawnCount;
@@ -66,7 +66,7 @@ public class EnemySpawner : MonoBehaviour
         enemy.gameObject.GetComponent<Enemy>().PeramPass(enemyTypes[enemySpawnType]);
     }
 
-     public void TakeDamage(int damage)
+     public void TakeDamage(float damage)
     {
         HP-= damage;
         StartCoroutine(DmgFlash());
